+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0]
 
-* TODO: Describe initial release.
+* First version
+* WebSocket and HTTPS (REST) communication
+* Only a few endpoints were tested